Put openpmd tests in the same file
